pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "contract-review"
version = "0.1.0"
description = "한국 법령 기반 계약서 자동 검토 CLI 도구"
readme = "README.md"
license = { text = "MIT" }
requires-python = ">=3.11"
keywords = ["contract", "review", "korean-law", "legal", "ai", "llm"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Legal Industry",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Office/Business",
]
dependencies = [
    "pydantic>=2.0",
    "pdfplumber>=0.11",
    "anthropic>=0.40",
    "google-generativeai>=0.8",
    "openai>=1.50",
    "typer>=0.12",
    "rich>=13",
    "python-dotenv>=1.0",
]

[project.scripts]
contract-review = "contract_review.cli:app"

[project.urls]
Homepage = "https://github.com/yourusername/contract-review"
Repository = "https://github.com/yourusername/contract-review"
Issues = "https://github.com/yourusername/contract-review/issues"

[tool.hatch.build.targets.wheel]
packages = ["src/contract_review"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]

[tool.ruff]
line-length = 100
target-version = "py311"
src = ["src"]

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W"]

[tool.mypy]
python_version = "3.11"
strict = true
mypy_path = "src"
